stats DB: Regroup SQL in maintenance/creation/use
SQL statements are grouped by their use: Generic schema independent
statements e.g. to start or finish a transaction are in group
"maintenance". All "create table" statements are in their own group. All
the data querying and updating stuff is in the third group.

This is useful, because different statements are needed at different
times. You can't e.g. prepare a data querying statement if the table
does not exist yet, or on the other hand prepare a table creation
statement if the table is already there.